Fix bug to not allow destroy of USTS Reg associated with a Race Reg
This commit fixes a bug where USTS Memberships could be deleted even if
there was an existing Race Registration tied to it. USTS Registrations
should not be able to be deleted if there is a Race Registration
associated with it. In the event of an override of that criteria, race
registrations will now be dependent destroyed along with the usts
registration.